Extracted code into 'AdaptorFactory'
